[SC] Update Instagram Social Info
