+++ src/constants/SocialUrls.js
@@ -2,4 +2,4 @@ export const TWITTER_URL = "https://twitter.com/spencerrc";
 export const LINKEDIN_URL = "https://www.linkedin.com/in/spencerrc";
 export const FACEBOOK_URL = "https://www.facebook.com/spencerrc";
 export const GITHUB_URL = "https://github.com/spencer-carver";
-export const INSTAGRAM_URL = "";
+export const INSTAGRAM_URL = "https://www.instagram.com/spencer_carver/";
